backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5

# Database
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.11.3

# API and serialization
marshmallow==3.20.1
webargs==8.3.0

# Authentication and security
PyJWT==2.8.0
passlib==1.7.4
email-validator==2.0.0

# PDF generation
reportlab==4.0.4

# File handling
python-magic==0.4.27
Pillow==10.0.0

# WebRTC and real-time
python-socketio==5.8.0
python-engineio==4.5.1
gevent==23.7.0
gevent-websocket==0.10.1

# Email
blinker==1.6.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0

# WSGI server
gunicorn==21.2.0